TPProfiling/Imgrot/requirements.txt
psutil
numpy
redis
pillow
opencv-python

